+++ packages/cli/loader/babel.js
@@ -1,11 +1,12 @@
 'use strict'
 
 const registerBabelLoader = require('../lib/babel')
+const context = require('../context')
 
 module.exports = function babelLoader(webpack) {
   registerBabelLoader(webpack.webpackChain, {
     name: 'javascript',
     test: [/\.js$/, /\.jsx$/],
-    include: webpack.paths.appPaths.contextPaths.concat(webpack.config.config.extraBabelCompileNodeModules),
+    include: context.paths.contextPaths.concat(context.config.config.extraBabelCompileNodeModules),
   })
 }